+++ packages/web/src/routes/form/success.svelte
@@ -7,7 +7,10 @@
 
   let emailInput: HTMLInputElement | undefined;
 
-  $: isValidMail = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/.test(email);
+  $: isValidMail =
+    /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/.test(
+      email
+    );
 
   async function handleEmailConfirm() {
     const response = await fetch(endpoint + "/api/add-email", {