+++ packages/web/src/types/metadata.ts
@@ -2,7 +2,8 @@ export type Metadata = {
   email: string | undefined;
   device: string;
   screenshot: string | undefined;
-  page: string;
+  siteUrl: string;
+  triggerElementText: string;
   // user defined metadata
   [userDefined: string]: string | undefined;
 };