+++ packages/web/CHANGELOG.md
@@ -1,4 +1,10 @@
+# 1.2.0
+
+- fix: previewing a screenshot
+- fix: email regex #2
+- feat: metadata.triggerElementText
+
 # 1.1.0
+
 - feat: loading animation while fetching config
 - feat: dark mode #1
-- fix: email regex #2